Stub out IP logging fields
I want to use Devise's `trackable` module to be able to track when users loast logged in, as I think that will be helpful in the future.

However, I don't need or want to be logging user's IP addresses. I feels a bit unnecessary and invasive. However Devise really wants to do this if trackable is on. So to bypass this, I'm adding these attr_accessors which allow devise to assign them, but they won't be persisted in the DB.