Custom registration process
We need to tweak the default devise registration process quite a bit:

We want the logged out home page to include the registration page, so I've created a partial that is used both on the home page, and the default devise location, so they are the same.

Although it's not the *most* secure, we want users to be logged in as soon as they confirm there email. It's secure enough for our needs.

We want to capture a user's name, and ensure they check the checkboxes, so add that to the form, and the permitted strong parameters

Install and use bootstrap_form to simplify form generation considerably.