Show buy ticket button if eventbrite event active
We use the eventbrite API to check if the event is live, and if so, we show the eventbrite modal ticket buying options.

Uses webmock to ensure we're not making actual API requests in the specs.
Add our first test that logs a user in, so create a helper method for that, which also uses factorybot for factories.

Add some README to explain how to setup the evntbrite EVN variables.