Create and set membership codes to users
We need all users to have a membership code.
These would normally be bulk generated before people start signing up so that we can export them to eventbrite before the event. However, if there aren't any available, we generate one instead of blowing up, and hopefully the admin will update Eventbrite.

I stole the code for CHARACTER_MAP from an old project of mine, which can make it easy to type membership codes back in while mapping confusing letters like 5 and S. We're only using it to generate the codes for now, so it's a bit overkill.