Support multiple events
Here is the minimum required to be able to support multiple events.

An event is stored in the DB with it's eventbrite credentials, instead of globally being set in the ENV.

There there is a lot of shifting things around in the REST model to support this.

This doesn't add any admin CRUD for events.