Make `active` event, and move that back to homepage
While supporting multiple events is probably a good idea for the future of the architecture, it's a bit weird adding an extra click when the current use case is probably only allowing one event at a time.

I've therefore moved all the event details back to the root path. And added a new boolean field to events to show which one is active.