Add mailchimp support
For London Decompression, we want to use Mailchimp to email members and ticket holders.

This commit submits the user's details to mailchimp after they have confirmed their email address is signup.

Adds two gems: Gibbon is for the Mailchimp API. ClimateControl allows us to set a mock ENV for Mailchimp in the tests.

Also update the README.