Add Low Income Requests
Based off of the last commit, the following feature have been added:

A member can make a single request for low income.
An admin can then view requests, and approve or reject them
If approved, we use the low income code instead of the membership code when redirecting to eventbrite.