Remove checking of eventbrite liveness
We've added the `active` field to events a few commit ago. This made the checking to see if the eventbrite event is live a bit pointless. We have an option to hide an event, so no need to rely on eventbrite to do this.

This simplifies the codebase, which is getting a bit overly complicated.