Create rake task to query eventbrite over tickets bought
This is a bit of a hack to support some last minute queries we want to do...

Because our membership system doesn't actually store whether a member has a ticket for an event or not (it pings the eventbrite API when that user makes a request), it makes it hard for us to create queries based on who has bought a ticket. We currently have two needs for this:

* Emailing all members who have bought a ticket but not volunteered, to encourage them to do so
* Email all successful low income ticket applicants who haven't bought a ticket yet to do so.

So this rake task goes through each user individually, and asks eventbrite if they have bought a ticket with their code, and just dumps that info on the user model. This will meet our needs for right now, but has two problems:

* It is going to be a slow rake task, that hits the eventbrite API quite hard
* This is only a snapshot, and the rake task will have to be run again if we want to update it
* Because I'm using the user model, it's based on the assumption that there's only one event, which won't be true next year