Check if tickets have sold out.
I wrote this to check if tickets have sold out for a given ticket class to give a better sold out message to members, while also still allowing low income tickets to purchase tickets.

I also re-used the `early_access` boolean to allow us to give access for directed sale. But I didn't think about the catch 22 here! If we're allowing direct access tickets, then we need to make sure that there are tickets available for that ticket class, which means they won't be sold out, so anyone can buy them! Doh.

So I've put this on a seperate branch in case I want to re-use this code in the future.