Add direct access code
For emergencies, we need to be able to give some members special access to buy a ticket, even if the main sale has sold out.

I've added some very quick support for this by creating another ticket class: 'DirectSale', which has it's own codes. Kind of similar to how Low Income is done.